Fix xattr support on linux
Remove support for xattr on symbolic links due to lack of syscall support in golang system package.

+++ hardlinks_unix.go
@@ -9,7 +9,7 @@ import (
 // hardlinkKey provides a tuple-key for managing hardlinks. This is system-
 // specific.
 type hardlinkKey struct {
-	dev   int32
+	dev   uint64
 	inode uint64
 }
